muserver-connectserver: refactoring.

+++ muserver-common/src/main/java/muserver/common/messages/PWMSG_HEAD2.java
@@ -9,7 +9,7 @@ import java.io.IOException;
 
 @AutoValue
 public abstract class PWMSG_HEAD2 extends AbstractPacket<PWMSG_HEAD2> {
- public static int sizeOf() {
+ public static byte sizeOf() {
   return 5;
  }
